# openapi/noteapi.yaml
openapi: 3.1.0
info:
  title: NoteAPI
  version: 0.1.0
servers:
  - url: https://noteapi.gould.pro
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    NoteFrontmatter:
      type: object
      properties: {}
      additionalProperties: true
...
paths:
...
  /notes:
    post:
      operationId: createNote
      summary: Create note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                frontmatter:
                  $ref: '#/components/schemas/NoteFrontmatter'
                content:
                  type: string
              required: [path]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
  /notes/{path}/move:
    post:
      operationId: moveNote
      summary: Move note
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPath:
                  type: string
              required: [newPath]
      responses:
        '200':
          description: Moved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...